Changed to reference new patch of everyauth
I have rejigged my everyauth patch to be on a branch with a large amount of mind-bending git-fu. This will enable us to track changes to the original project more effectively and is the recommended approach. In the process I have rebased my changes onto the latest version of everyauth, so any fixes he has made recently should be in there. I'm still hopeful that he will merge my pull request or something like it in the very near future.